+++ Sources/APRelayCore/RelayConfiguration.swift
@@ -11,6 +11,7 @@ package struct RelayConfiguration: Sendable {
     package let adminToken: String
     package let manualAccept: Bool
     package let restrictedMode: Bool
+    package let nodeInfoCheckInterval: Int
     package let relayName: LocalizedString
     package let relayDescription: LocalizedString
     package let relayFooter: LocalizedString
@@ -29,6 +30,7 @@ package struct RelayConfiguration: Sendable {
         adminToken: String = "",
         manualAccept: Bool = false,
         restrictedMode: Bool = false,
+        nodeInfoCheckInterval: Int = 300,
         relayName: LocalizedString = LocalizedString([:]),
         relayDescription: LocalizedString = LocalizedString([:]),
         relayFooter: LocalizedString = LocalizedString([:])
@@ -43,6 +45,7 @@ package struct RelayConfiguration: Sendable {
         self.adminToken = adminToken
         self.manualAccept = manualAccept
         self.restrictedMode = restrictedMode
+        self.nodeInfoCheckInterval = nodeInfoCheckInterval
         self.relayName = relayName
         self.relayDescription = relayDescription
         self.relayFooter = relayFooter